landing page con nav reutilizable
