+++ client/src/components/NavLanding/NavLanding.jsx
@@ -1,37 +1,12 @@
-import React, { useEffect } from "react";
-import { useState } from "react";
 import { Link } from "react-router-dom";
 import logofusionajob from '../../assets/logofusionajob.png'
-import dia from '../../assets/sun.png'
-import noche from '../../assets/moon.png'
 import perfil from '../../assets/user.png'
-import { ModalLogin } from "../ModalLogin/ModalLogin"
+import ModoNoche from "./ModoNoche";
+import ProfileButton from "./ProfileButton";
 
-export const NavLanding = () => {
-    const localDark = JSON.parse(localStorage.getItem('isDarkMode')) || false
-    const [isDarkMode, setIsDarkMode] = useState(localDark);
-
-    const handleToggle = () => {
-      const newIsDarkMode = !isDarkMode;
-      setIsDarkMode(newIsDarkMode);
-      if (!newIsDarkMode) {
-        document.documentElement.classList.remove('dark');
-        localStorage.setItem('isDarkMode', 'false');
-      } else {
-        document.documentElement.classList.add('dark');
-        localStorage.setItem('isDarkMode', 'true');
-      }
-    };
-
-    useEffect(() => {
-      if(localDark) document.documentElement.classList.add('dark')
-      else document.documentElement.classList.remove('dark')
-    }, [localDark])
-    
-    const [open, setOpen] = useState(false);
-
+export const NavLanding = ({menu}) => {
     return (
-    <nav className="bg-secondary-light dark:bg-primary-dark fixed w-full transition-all">
+    <nav className="bg-secondary-light dark:bg-primary-dark fixed top-0 w-full transition-all">
       <div className="max-w-7xl mx-auto px-4 sm:px-6 lg:px-8">
         <div className="flex items-center justify-between h-16">
           {/* Logo */}
@@ -43,30 +18,25 @@ export const NavLanding = () => {
             />
           </div>
           {/* Menú */}
-          <div className="hidden md:block">
+          <div className="block">
             <div className="ml-6 flex items-baseline space-x-4 dark:text-text-dark text-gray-900">
-            <Link className="dark:hover:text-yellow-200 hover:text-xl transition-all inline-block font-medium px-4 py-2">Planes</Link>
-            <Link to='/about' className="dark:hover:text-yellow-200 hover:text-xl transition-all inline-block font-medium px-4 py-2">Sobre nosotros</Link>
-            <Link to='/dashboardempresa' className="dark:hover:text-yellow-200 hover:text-xl transition-all inline-block font-medium px-4 py-2">Dashboard Empresa</Link>
-            <Link to='/registro' className="dark:hover:text-yellow-200 hover:text-xl transition-all inline-block font-medium px-4 py-2">Registro</Link>
-          </div>
+            {
+              menu?.map(item => {
+                return <Link key={item.name} to={item.link} className="dark:hover:text-yellow-200 hover:text-xl transition-all inline-block font-medium px-4 py-2">{ item.name }</Link>
+              })
+            }
+            </div>
           </div>
           {/* Botones */}
-          <div className="hidden md:block">
+          <div className="block">
             <div className="ml-4 flex items-center md:ml-6 gap-3">
-            <button onClick={() => {setOpen(!open)}} className="py-2 px-2 bg-gray-300 text-black dark:bg-slate-500 dark:text-white font-semibold rounded-lg shadow-md hover:bg-gray-400 focus:outline-none focus:ring-2">
-            <img className="w-4 inline-block align-middle mr-2" src={perfil} alt='ingresar'/>
-              Ingresar
-            </button>
-            <div onClick={handleToggle} className="cursor-pointer py-2 px-2 bg-gray-300 rounded-lg shadow-md hover:bg-gray-400 dark:hover:bg-secondary-light">{isDarkMode 
-            ? <img className="w-6" src={dia} alt='dia'/>
-            : <img className="w-6" src={noche} alt='noche'/>
-            }</div>
-              </div>
+            <ProfileButton perfil={perfil} />
+            <ModoNoche />
+            </div>
           </div>
         </div>
       </div>
-      <ModalLogin isOpen={open} setOpen={setOpen}/>
+            
     </nav>
 )
 }
